split code for readability, refactor some parts
